backend/assistant/requirements.txt
crewai[tools]
pdf2image
pdfplumber
pi-heif
pillow
pymupdf
pymupdf4llm
pypdf
redis
unstructured
unstructured-inference
unstructured-pytesseract
qdrant-client
ollama
transformers
sentence-transformers
torch
fastapi
uvicorn[standard]
streamlit
requests
pyrefly
clip
openai-clip
python-dotenv
websockets
pyaudio

# Video Report Dependencies
opencv-python>=4.8.0
moviepy>=1.0.3
markdown>=3.4.0
httpx>=0.24.0
groq>=0.4.0
python-multipart>=0.0.6
aiofiles>=23.0.0
librosa>=0.10.0
scipy>=1.10.0
torchaudio>=2.0.0
transformers>=4.30.0
